+++ relayer-cli/src/commands/keys/restore.rs
@@ -55,9 +55,7 @@ impl Runnable for KeyRestoreCmd {
         let config = app_config();
 
         let opts = match self.validate_options(&config) {
-            Err(err) => {
-                return Output::error(err).exit();
-            }
+            Err(err) => return Output::error(err).exit(),
             Ok(result) => result,
         };