+++ relayer-cli/src/commands/tx/channel.rs
@@ -5,16 +5,17 @@ use ibc::ics04_channel::channel::Order;
 use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
 use ibc::Height;
 use relayer::channel::{Channel, ChannelSide};
+use relayer::config::StoreConfig;
 
-use crate::commands::cli_utils::chain_handlers_from_chain_id;
+use crate::commands::cli_utils::{ChainHandlePair, SpawnOptions};
 use crate::conclude::Output;
 use crate::error::{Error, Kind};
 use crate::prelude::*;
 
-macro_rules! chan_open_cmd {
-    ($chan_open_cmd:ident, $dbg_string:literal, $func:ident) => {
+macro_rules! tx_chan_cmd {
+    ($tx_chan_cmd:ident, $dbg_string:literal, $func:ident) => {
         #[derive(Clone, Command, Debug, Options)]
-        pub struct $chan_open_cmd {
+        pub struct $tx_chan_cmd {
             #[options(free, required, help = "identifier of the destination chain")]
             dst_chain_id: ChainId,
 
@@ -36,23 +37,26 @@ macro_rules! chan_open_cmd {
             #[options(free, required, help = "identifier of the source channel")]
             src_channel_id: ChannelId,
 
-            #[options(help = "the channel order", short = "o")]
+            #[options(
+                help = "the channel order: `UNORDERED` or `ORDERED`, default `UNORDERED`",
+                short = "o"
+            )]
             ordering: Order,
         }
 
-        impl Runnable for $chan_open_cmd {
+        impl Runnable for $tx_chan_cmd {
             fn run(&self) {
                 let config = app_config();
 
-                let chains = match chain_handlers_from_chain_id(
+                let spawn_options = SpawnOptions::override_store_config(StoreConfig::memory());
+                let chains = match ChainHandlePair::spawn_with(
+                    spawn_options,
                     &config,
                     &self.src_chain_id,
                     &self.dst_chain_id,
                 ) {
                     Ok(chains) => chains,
-                    Err(e) => {
-                        return Output::error(format!("{}", e)).exit();
-                    }
+                    Err(e) => return Output::error(format!("{}", e)).exit(),
                 };
 
                 // Retrieve the connection
@@ -86,21 +90,45 @@ macro_rules! chan_open_cmd {
 
                 match res {
                     Ok(receipt) => Output::success(receipt).exit(),
-                    Err(e) => Output::error(format!("{}", e)).exit(),
+                    Err(e) => Output::error(format!("{:?}", e)).exit(),
                 }
             }
         }
     };
 }
 
-chan_open_cmd!(TxRawChanInitCmd, "ChanOpenInit", build_chan_init_and_send);
+tx_chan_cmd!(
+    TxRawChanOpenInitCmd,
+    "ChanOpenInit",
+    build_chan_open_init_and_send
+);
 
-chan_open_cmd!(TxRawChanTryCmd, "ChanOpenTry", build_chan_try_and_send);
+tx_chan_cmd!(
+    TxRawChanOpenTryCmd,
+    "ChanOpenTry",
+    build_chan_open_try_and_send
+);
 
-chan_open_cmd!(TxRawChanAckCmd, "ChanOpenAck", build_chan_ack_and_send);
+tx_chan_cmd!(
+    TxRawChanOpenAckCmd,
+    "ChanOpenAck",
+    build_chan_open_ack_and_send
+);
 
-chan_open_cmd!(
-    TxRawChanConfirmCmd,
+tx_chan_cmd!(
+    TxRawChanOpenConfirmCmd,
     "ChanOpenConfirm",
-    build_chan_confirm_and_send
+    build_chan_open_confirm_and_send
+);
+
+tx_chan_cmd!(
+    TxRawChanCloseInitCmd,
+    "ChanCloseInit",
+    build_chan_close_init_and_send
+);
+
+tx_chan_cmd!(
+    TxRawChanCloseConfirmCmd,
+    "ChanCloseConfirm",
+    build_chan_close_confirm_and_send
 );