+++ relayer/src/chain.rs
@@ -1,42 +1,35 @@
-pub(crate) mod cosmos;
-pub use cosmos::CosmosSDKChain;
-
-pub mod handle;
-pub mod runtime;
-
-#[cfg(test)]
-pub mod mock;
+use std::{sync::Arc, thread};
 
 use crossbeam_channel as channel;
-use std::{sync::Arc, thread};
-use tokio::runtime::Runtime as TokioRuntime;
-
 use prost_types::Any;
-
-use tendermint_proto::Protobuf;
-
 // TODO - tendermint deps should not be here
 use tendermint::account::Id as AccountId;
 use tendermint::block::Height;
+use tendermint_proto::Protobuf;
+use tokio::runtime::Runtime as TokioRuntime;
 
-use ibc_proto::ibc::core::channel::v1::{
-    PacketState, QueryConnectionChannelsRequest, QueryPacketAcknowledgementsRequest,
-    QueryPacketCommitmentsRequest, QueryUnreceivedAcksRequest, QueryUnreceivedPacketsRequest,
-};
-use ibc_proto::ibc::core::client::v1::QueryClientStatesRequest;
-use ibc_proto::ibc::core::commitment::v1::MerkleProof;
-
+pub use cosmos::CosmosSDKChain;
 use ibc::events::IBCEvent;
 use ibc::ics02_client::header::Header;
 use ibc::ics02_client::state::{ClientState, ConsensusState};
 use ibc::ics03_connection::connection::{ConnectionEnd, State};
+use ibc::ics03_connection::raw::ConnectionIds;
 use ibc::ics03_connection::version::{get_compatible_versions, Version};
 use ibc::ics04_channel::channel::{ChannelEnd, QueryPacketEventDataRequest};
+use ibc::ics04_channel::packet::{PacketMsgType, Sequence};
 use ibc::ics23_commitment::commitment::{CommitmentPrefix, CommitmentProofBytes};
 use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
 use ibc::ics24_host::Path;
 use ibc::proofs::{ConsensusProof, Proofs};
 use ibc::Height as ICSHeight;
+use ibc_proto::ibc::core::channel::v1::{
+    PacketState, QueryChannelsRequest, QueryConnectionChannelsRequest,
+    QueryPacketAcknowledgementsRequest, QueryPacketCommitmentsRequest, QueryUnreceivedAcksRequest,
+    QueryUnreceivedPacketsRequest,
+};
+use ibc_proto::ibc::core::client::v1::QueryClientStatesRequest;
+use ibc_proto::ibc::core::commitment::v1::MerkleProof;
+use ibc_proto::ibc::core::connection::v1::QueryConnectionsRequest;
 
 use crate::config::ChainConfig;
 use crate::connection::ConnectionMsgType;
@@ -44,7 +37,13 @@ use crate::error::{Error, Kind};
 use crate::event::monitor::EventBatch;
 use crate::keyring::store::{KeyEntry, KeyRing};
 use crate::light_client::LightClient;
-use ibc::ics04_channel::packet::{PacketMsgType, Sequence};
+
+pub(crate) mod cosmos;
+pub mod handle;
+pub mod runtime;
+
+#[cfg(test)]
+pub mod mock;
 
 /// Generic query response type
 /// TODO - will slowly move to GRPC protobuf specs for queries
@@ -304,6 +303,7 @@ pub trait Chain: Sized {
                 CommitmentProofBytes::from(connection_proof),
                 client_proof,
                 consensus_proof,
+                None,
                 height.increment(),
             )
             .map_err(|_| Kind::MalformedProof)?,
@@ -332,8 +332,10 @@ pub trait Chain: Sized {
         let channel_proof =
             CommitmentProofBytes::from(self.proven_channel(port_id, channel_id, height)?.1);
 
-        Ok(Proofs::new(channel_proof, None, None, height.increment())
-            .map_err(|_| Kind::MalformedProof)?)
+        Ok(
+            Proofs::new(channel_proof, None, None, None, height.increment())
+                .map_err(|_| Kind::MalformedProof)?,
+        )
     }
 
     fn query_packet_commitments(
@@ -365,6 +367,12 @@ pub trait Chain: Sized {
     /// Performs a query to retrieve the identifiers of all clients associated with a chain.
     fn query_clients(&self, request: QueryClientStatesRequest) -> Result<Vec<ClientId>, Error>;
 
+    /// Performs a query to retrieve the identifiers of all connections.
+    fn query_connections(&self, request: QueryConnectionsRequest) -> Result<ConnectionIds, Error>;
+
+    /// Performs a query to retrieve the identifiers of all channels.
+    fn query_channels(&self, request: QueryChannelsRequest) -> Result<Vec<ChannelId>, Error>;
+
     fn build_packet_proofs(
         &self,
         packet_type: PacketMsgType,
@@ -374,6 +382,8 @@ pub trait Chain: Sized {
         height: ICSHeight,
     ) -> Result<(Vec<u8>, Proofs), Error> {
         let data: Path;
+        let mut channel_proof = None;
+
         match packet_type {
             PacketMsgType::Recv => {
                 data = Path::Commitments {
@@ -396,6 +406,16 @@ pub trait Chain: Sized {
                     sequence,
                 }
             }
+            PacketMsgType::TimeoutOnClose => {
+                data = Path::Receipts {
+                    port_id: port_id.clone(),
+                    channel_id: channel_id.clone(),
+                    sequence,
+                };
+                channel_proof = Some(CommitmentProofBytes::from(
+                    self.proven_channel(port_id, channel_id, height)?.1,
+                ));
+            }
         }
 
         let res = self
@@ -408,6 +428,7 @@ pub trait Chain: Sized {
             })?),
             None,
             None,
+            channel_proof,
             height.increment(),
         )
         .map_err(|_| Kind::MalformedProof)?;