+++ relayer/src/relay.rs
@@ -28,9 +28,19 @@ pub fn relay_on_new_link(
     Ok(())
 }
 
-/// TODO - CLI needs to be done, parameters TBD
+/// Relays packets over a specified channel
+/// Used by the `rrly -c config.toml start ibc-0 ibc-1 transfer channel-0`
+pub fn channel_relay(
+    a_chain: Box<dyn ChainHandle>,
+    b_chain: Box<dyn ChainHandle>,
+    opts: &LinkParameters,
+) -> Result<(), BoxError> {
+    let mut link = Link::new_from_opts(a_chain, b_chain, opts)?;
+    Ok(link.relay()?)
+}
+
 /// Connects two ports of two chains creating new clients, connection and channel
-/// Used by the `rrly -c config.toml relay channel init ibc-0 ibc-1 transfer transfer `
+/// Used by the `rrly -c config.toml channel handshake ibc-0 ibc-1 transfer transfer `
 pub fn connect_with_new_channel(
     a_chain_handle: Box<dyn ChainHandle>,
     b_chain_handle: Box<dyn ChainHandle>,
@@ -54,19 +64,3 @@ pub fn connect_with_new_channel(
         path.b_port,
     )?)
 }
-
-/// TODO - CLI needs to be done, parameters TBD
-/// Relays packets over a specified channel
-/// Used by the `rrly -c config.toml start ibc-0 ibc-1 transfer channel-0`
-pub fn channel_relay(
-    a_chain: Box<dyn ChainHandle>,
-    b_chain: Box<dyn ChainHandle>,
-    opts: &LinkParameters,
-) -> Result<(), BoxError> {
-    let mut link = Link::new_from_opts(a_chain, b_chain, opts)?;
-    Ok(link.relay()?)
-}
-
-// rrly start -> handshakes + packet relay
-// rrly init -> handshakes
-// rrly start ibc-0 ibc-1 transfer channel-0 -> packet relay on specified channel